+++ src/evaluation/model/fcn_model.py
@@ -3,18 +3,10 @@ import torch
 
 
 class Model(nn.Module):
-    def __init__(self, input_dim, conditional_dim, hidden_dim, conditional, num_layers=2):
+    def __init__(self, input_dim):
         super().__init__()
-        self.lstm = nn.LSTM(input_size=input_dim, hidden_size=hidden_dim, num_layers=2, dropout=0.3)
-        self.conditional = conditional
-        if self.conditional:
-            self.fc1 = nn.Sequential(
-            nn.Linear(in_features=hidden_dim+conditional_dim, out_features=64),
-            nn.ReLU()
-        )
-        else:
-            self.fc1 = nn.Sequential(
-            nn.Linear(in_features=hidden_dim, out_features=64),
+        self.fc1 = nn.Sequential(
+            nn.Linear(in_features=input_dim, out_features=64),
             nn.ReLU()
         )
 
@@ -29,14 +21,8 @@ class Model(nn.Module):
         
         self.fc4 = nn.Linear(16, 1)
         
-    def forward(self, x, c):
-        out, (h,d) = self.lstm(x.unsqueeze(1))
-        out = out.squeeze(1)
-
-        if self.conditional:
-            out = torch.cat([out,c], dim=1)
-        
-        y = self.fc1(out)
+    def forward(self, x):
+        y = self.fc1(x)
         y = self.fc2(y)
         y = self.fc3(y)
         y = self.fc4(y)
